requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-huggingface==0.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
langchain-community==0.0.10